상품 서비스 개발

+++ jpashop_webver/src/main/java/com/webver/jpabook/jpashop_webver/repository/ItemRepository.java
@@ -33,7 +33,7 @@ public class ItemRepository {
      * @param itemId
      * @return
      */
-    public Item findOne(int itemId) {
+    public Item findOne(Long itemId) {
         return em.find(Item.class, itemId);
     }
